Improve support for SPI
Motivation:

Current use of SPI is quite limited: it is missing a mechanism to
select which plugin to use, to configure that plugin, or the ability to
use all of the functionality of that plugin.

Modification:

Add a factory interface that forms the mechanism for choosing which
plugin is needed and takes care of instantiating that plugin.  Also
added a mechanism to select a plugin and access all the functionality
of the SPI.

Result:

Better support for selecting and using a QoS plugin.