Add Subject to all plugin operations.
Motivation:

The plugin may wish to make authorisation decisions based on the user's
identity.

Modification:

Add a Subject (from JAAS) as an argument to each method.  This allows
the front-end to include whichever principals identify the user.

Result:

The plugin is provided with user-identity information, so can choose
whether or not to accept a request.