Add support for discovering the users identity
Motivation:

The plugin is responsible for authorisation decisions.  To achieve this,
it needs to know who is the user.

Modification:

Add support for providing the user's identity.  This is achieved using
the concepts from JAAS: a Subject is provided as a thread-local value
that the plugin can discover as necessary.  Information about the user
is available as principals (such as the username) or as public
credentials (such as OIDC token).

Result:

The plugin should have sufficient information about the user to make
authorisation decisions.