Add missing exception from StorageBackend methods
Motivation:

In general, all queries to a plugin may fail; for example, the plugin
may loose connection with the backend system, or a user is not
authorised for any activity or discovering anything about the current
state of the backend system.

Modification:

Add exceptions to getCapabilities and getCurrentStatus methods.
Related classes are updated accordingly.

Result:

All methods provide the plugin with the possibility to report that the
operation failed.